return-it-app/backend/requirements.txt
fastapi
uvicorn
opencv-python
google-cloud-vision
google-auth
google-auth-oauthlib
pillow
